% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{at_get_history}
\alias{at_get_history}
\title{Get transformation history}
\usage{
at_get_history(data)
}
\arguments{
\item{data}{Either a data frame returned by at_transform() or a result
object from at_transform(verbose = TRUE)}
}
\value{
Data frame summarizing transformations
}
\description{
Get transformation history
}
\examples{
\dontrun{
# Works with both verbose=FALSE and verbose=TRUE
transformed_df <- at_transform(df)
history <- at_get_history(transformed_df)
}
}
