% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{autotransform}
\alias{autotransform-package}
\alias{autotransform}
\title{Automatic Data Type Detection and Transformation with Rollback}
\description{
Automatically detects and transforms ambiguous data types in
data frames. Handles numeric strings, categorical
variables, and more. Includes rollback functionality to revert
transformations entirely or for specific columns. Perfect for cleaning
messy datasets with mixed or unclear data types.
}
\author{
Abednego Nasila \href{mailto:alisanabednego@gmail.com}{alisanabednego@gmail.com}
}
\keyword{internal}
