% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{convert_na_strings}
\alias{convert_na_strings}
\title{Convert N/A strings to R's NA}
\usage{
convert_na_strings(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Data frame with N/A strings converted to NA
}
\description{
Convert N/A strings to R's NA
}
\keyword{internal}
