% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{determine_data_type}
\alias{determine_data_type}
\title{Determine best data type for a column}
\usage{
determine_data_type(x, col_name = "")
}
\arguments{
\item{x}{A vector}

\item{col_name}{Column name for logging (optional)}
}
\value{
Character string indicating recommended type
}
\description{
Determine best data type for a column
}
\keyword{internal}
