% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{is_mostly_date}
\alias{is_mostly_date}
\title{Check if column contains dates}
\usage{
is_mostly_date(x, threshold = 0.6)
}
\arguments{
\item{x}{A vector}

\item{threshold}{Proportion threshold (default: 0.6)}
}
\value{
Logical
}
\description{
Check if column contains dates
}
