% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{is_mostly_factor}
\alias{is_mostly_factor}
\title{Check if column is mostly categorical/factor}
\usage{
is_mostly_factor(x, max_unique_ratio = 0.5, min_frequency = 2)
}
\arguments{
\item{x}{A vector}

\item{max_unique_ratio}{Maximum ratio of unique to total values (default: 0.5)}

\item{min_frequency}{Minimum frequency for most common value (default: 2)}
}
\value{
Logical
}
\description{
Check if column is mostly categorical/factor
}
