% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{is_mostly_logical}
\alias{is_mostly_logical}
\title{Check if column is mostly logical/boolean}
\usage{
is_mostly_logical(x, threshold = 0.8)
}
\arguments{
\item{x}{A vector}

\item{threshold}{Proportion threshold (default: 0.8)}
}
\value{
Logical
}
\description{
Check if column is mostly logical/boolean
}
