% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{is_mostly_numeric}
\alias{is_mostly_numeric}
\title{Check if column is mostly numeric}
\usage{
is_mostly_numeric(x, threshold = 0.8)
}
\arguments{
\item{x}{A vector}

\item{threshold}{Proportion threshold (default: 0.8)}
}
\value{
Logical
}
\description{
Check if column is mostly numeric
}
