% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{print.at_data}
\alias{print.at_data}
\title{Print method for at_data objects}
\usage{
\method{print}{at_data}(x, ...)
}
\arguments{
\item{x}{A at_data object}

\item{...}{Additional arguments passed to default print}
}
\description{
Print method for at_data objects
}
