% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{print.at_transform}
\alias{print.at_transform}
\title{Print method for at_transform objects}
\usage{
\method{print}{at_transform}(x, ...)
}
\arguments{
\item{x}{A at_transform object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for at_transform objects
}
