% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{safe_convert}
\alias{safe_convert}
\title{Safely convert data types}
\usage{
safe_convert(x, target_type, col_name = "")
}
\arguments{
\item{x}{A vector to convert}

\item{target_type}{Target data type}

\item{col_name}{Column name for logging}
}
\value{
Converted vector
}
\description{
Safely convert data types
}
\keyword{internal}
